@@ -1,7 +1,7 @@
 {
   "name": "@contentstack/cli-cm-regex-validate",
   "description": "Validate Fields with Regex Property of Content Type and Global Field in a Stack",
-  "version": "1.0.1",
+  "version": "1.0.2",
   "author": "Contentstack",
   "bugs": "https://github.com/contentstack/cli-cm-regex-validate/issues",
   "devDependencies": {
@@ -11,21 +11,21 @@
     "@types/chai": "^4.2.21",
     "@types/configstore": "^5.0.1",
     "@types/inquirer": "^8.1.1",
-    "@types/jest": "^27.0.1",
+    "@types/jest": "^29.1.2",
     "@types/jsonexport": "^3.0.2",
-    "@types/mocha": "^5.2.7",
+    "@types/mocha": "^10.0.0",
     "@types/node": "^10.17.60",
     "@types/safe-regex": "^1.1.3",
     "chai": "^4.3.4",
-    "eslint": "^5.16.0",
-    "eslint-config-oclif": "^3.1.0",
-    "eslint-config-oclif-typescript": "^0.1.0",
+    "eslint": "^8.25.0",
+    "eslint-config-oclif": "^4.0.0",
+    "eslint-config-oclif-typescript": "^1.0.3",
     "globby": "^10.0.2",
-    "mocha": "^5.2.0",
-    "nyc": "^14.1.1",
-    "ts-jest": "^27.0.5",
-    "ts-node": "^8.10.2",
-    "typescript": "^3.9.10"
+    "mocha": "^10.0.0",
+    "nyc": "^15.1.0",
+    "ts-jest": "^29.0.3",
+    "ts-node": "^10.9.1",
+    "typescript": "^4.8.4"
   },
   "engines": {
     "node": ">=8.0.0"
@@ -62,14 +62,15 @@
     "version": "oclif-dev readme && git add README.md"
   },
   "dependencies": {
-    "@contentstack/cli-command": "^0.1.1-beta.2",
+    "@contentstack/cli-command": "^1.0.0",
     "@contentstack/management": "^1.2.4",
+    "@contentstack/cli-utilities": "^1.0.0",
     "@oclif/command": "^1.8.0",
     "@oclif/config": "^1.17.0",
     "cli-table3": "^0.6.0",
-    "cli-ux": "^5.6.3",
+    "cli-ux": "^6.0.9",
     "inquirer": "^6.5.2",
-    "jest": "^27.1.1",
+    "jest": "^29.1.2",
     "jsonexport": "^3.2.0",
     "safe-regex": "^1.1.0",
     "tslib": "^1.14.1"
